requirements.txt
streamlit
tensorflow
numpy
Pillow
fastapi
uvicorn[standard]
python-multipart
